+++ CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Nothing
 
+## [0.6.0] - 2017-01-26
+
+### Added
+- Added longtask endpoints
+- A markdown file containing all endpoints with their current state
+
+### Changed
+- client.spaces is renamed to client.get_spaces in keeping with other endpoints
+
 ## [0.5.0] - 2017-01-26
 
 ### Added
@@ -56,7 +65,8 @@ First public release of the library
 ### Changed
 - Nothing
 
-[Unreleased]: https://github.com/DaveTCode/confluence-python-lib/compare/0.5.0...HEAD
+[Unreleased]: https://github.com/DaveTCode/confluence-python-lib/compare/0.6.0...HEAD
+[0.5.0]: https://github.com/DaveTCode/confluence-python-lib/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/DaveTCode/confluence-python-lib/compare/0.3.0...0.5.0
 [0.3.0]: https://github.com/DaveTCode/confluence-python-lib/compare/0.2.2...0.3.0
 [0.2.1]: https://github.com/DaveTCode/confluence-python-lib/compare/0.2.1...0.2.2